refactor(Images): Split `ImageReference` into own module
