+++ rust/images/src/lib.rs
@@ -2,6 +2,7 @@ pub mod blob_writer;
 pub mod change_set;
 pub mod distribution;
 pub mod image;
+pub mod image_reference;
 pub mod media_types;
 pub mod snapshot;
 mod utils;
